Chore: add more strict rules to PHPStan (#82)
* Add phpstan extensions

* Replace `?:` in favor of `??`

* Fixes

* We don't need this test because will always be true

* fixes

+++ tests/Example/DemoTest.php
@@ -122,7 +122,7 @@ class DemoTest extends TestCase
 
     private function assertSnapshot(string $name, DummyBackend $backend): void
     {
-        $name = substr($name, strrpos($name, ':') + 5);
+        $name = substr($name, (int) strrpos($name, ':') + 5);
         $filename = sprintf('%s/snapshot/%s.snapshot', __DIR__, $name);
         if (!file_exists($filename) || getenv('SNAPSHOT_APPROVE')) {
             file_put_contents($filename, $backend->flushed());
